+++ src/xrGame/xrServer.cpp
@@ -826,7 +826,7 @@ void xrServer::Server_Client_Check(IClient* CL)
         return;
     };
 
-    if (CL->process_id == GetCurrentProcessId())
+    if (static_cast<int>(CL->process_id) == GetCurrentProcessId())
     {
         CL->flags.bLocal = 1;
         SV_Client = (xrClientData*)CL;
